+++ src/PrestaShopBundle/Resources/config/services/adapter/product.yml
@@ -233,7 +233,7 @@ services:
     prestashop.adapter.product.command_handler.remove_all_associated_product_suppliers_handler:
         class: PrestaShop\PrestaShop\Adapter\Product\CommandHandler\RemoveAllAssociatedProductSuppliersHandler
         arguments:
-            - '@prestashop.adapter.product.product_supplier_repository'
+            - '@prestashop.adapter.product.repository.product_supplier_repository'
             - '@prestashop.adapter.product.product_supplier_updater'
         tags:
             - name: tactician.handler
@@ -319,8 +319,8 @@ services:
             - '@prestashop.adapter.supplier.supplier_repository'
             - '@prestashop.adapter.currency.currency_repository'
 
-    prestashop.adapter.product.product_supplier_repository:
-        class: PrestaShop\PrestaShop\Adapter\Product\ProductSupplierRepository
+    prestashop.adapter.product.repository.product_supplier_repository:
+        class: PrestaShop\PrestaShop\Adapter\Product\Repository\ProductSupplierRepository
         arguments:
             - '@prestashop.adapter.product.product_supplier_validator'
 
@@ -329,4 +329,4 @@ services:
         arguments:
             - '@prestashop.adapter.product.product_repository'
             - '@prestashop.adapter.supplier.supplier_repository'
-            - '@prestashop.adapter.product.product_supplier_repository'
+            - '@prestashop.adapter.product.repository.product_supplier_repository'
